Fixed new spdlog formatting - temp fix

+++ engine/include/real/api/gl/gl_conversions.hpp
@@ -30,7 +30,7 @@ namespace Real
 			case shader_data_t::bvec4 : return GL_BOOL;
 
 			default:
-			case shader_data_t::none: REAL_CORE_ERROR("Unsupported data type: {}!", type);
+            case shader_data_t::none: REAL_CORE_ERROR("Unsupported data type: {}!", static_cast<int32_t>(type));
 				return 0;
 		}
 		// @formatter:on
