+++ engine/include/real/renderer/buffer_layout.hpp
@@ -59,9 +59,7 @@ namespace Real
 				case shader_data_t::bvec4 : return 1 * 4;
 
 				default:
-				case shader_data_t::none:
-					REAL_CORE_ERROR("Unsupported data type: {}!",
-							type);
+                case shader_data_t::none: REAL_CORE_ERROR("Unsupported data type: {}!", static_cast<int32_t>(type));
 					return 0;
 			}
 			// @formatter:on
@@ -77,8 +75,8 @@ namespace Real
 
 			default:
 			case RendererAPI::API::None:
-				REAL_CORE_ERROR("Invalid renderer api: {}",
-						RendererAPI::API::None);
+                REAL_CORE_ERROR("Invalid renderer api: {}",
+                                static_cast<int32_t>(RendererAPI::API::None));
 				return -1;
 			}
 		}
