+++ engine/src/renderer/array_vertex.cpp
@@ -18,7 +18,7 @@ namespace Real
 		default:
 		case RendererAPI::API::None:
 			REAL_CORE_ERROR("Invalid renderer api: {}",
-					RendererAPI::API::None);
+					static_cast<int32_t>(RendererAPI::API::None));
 			return nullptr;
 		}
 	}
